requirements.txt
PyQt6==6.6.1
PyQt6-sip==13.6.0
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.0
pynput==1.7.6
pyperclip==1.8.2
cryptography==44.0.2
